requirements.txt
pandas==2.3.3
openpyxl==3.1.2
python-Levenshtein==0.27.3
pytest==9.0.1
requests==2.31.0
